fix: "Implicit conversion loses integer precision" warnings (#3960)

+++ macosx/PrefsController.h
@@ -11,7 +11,8 @@
 
 @property(nonatomic, readonly) NSArray<NSString*>* sounds;
 
-+ (NSInteger)dateToTimeSum:(NSDate*)date;
+/// - returns: number of minutes
++ (int)dateToTimeSum:(NSDate*)date;
 
 - (instancetype)initWithHandle:(tr_session*)handle;
